+++ src/fsw/FCCode/DownlinkProducer.cpp
@@ -31,7 +31,7 @@ void DownlinkProducer::init_flows(const std::vector<FlowData>& flow_data) {
     // Create flow objects out of the flow data. Ensure that
     // no two flows have the same ID.
     std::set<unsigned char> ids;
-    const size_t num_flows = flow_data.size();
+    const unsigned int num_flows = flow_data.size();
     flows.reserve(num_flows);
     for (const FlowData& flow : flow_data) {
         if (ids.find(flow.id) != ids.end()) {
@@ -45,14 +45,14 @@ void DownlinkProducer::init_flows(const std::vector<FlowData>& flow_data) {
     // Set the snapshot size to the maximum possible downlink size,
     // so that the Quake Manager allocates sufficient space for
     // its copy of the snapshot.
-    const size_t max_downlink_size = compute_max_downlink_size();
+    const unsigned int max_downlink_size = compute_max_downlink_size();
     snapshot = new char[max_downlink_size];
     snapshot_ptr_f.set(snapshot);
     snapshot_size_bytes_f.set(max_downlink_size);
 }
 
-size_t DownlinkProducer::compute_downlink_size(const bool compute_max) const {
-    size_t downlink_max_size_bits = 0;
+unsigned int DownlinkProducer::compute_downlink_size(const bool compute_max) const {
+    unsigned int downlink_max_size_bits = 0;
 
     for (const Flow& flow : flows) {
         if (flow.is_active || compute_max)
@@ -66,21 +66,21 @@ size_t DownlinkProducer::compute_downlink_size(const bool compute_max) const {
     downlink_max_size_bits += (downlink_max_size_bits + num_bits_in_packet - 1) / num_bits_in_packet;
 
     // Byte-align downlink packet
-    const size_t downlink_max_size_bytes = (downlink_max_size_bits + 7) / 8;
+    const unsigned int downlink_max_size_bytes = (downlink_max_size_bits + 7) / 8;
 
     return downlink_max_size_bytes;
 }
 
-size_t DownlinkProducer::compute_max_downlink_size() const {
+unsigned int DownlinkProducer::compute_max_downlink_size() const {
     return compute_downlink_size(true);
 }
 
 static void add_bits_to_downlink_frame(const bit_array& field_bits,
                                        char* snapshot_ptr,
-                                       size_t& packet_offset,
-                                       size_t& downlink_frame_offset)
+                                       unsigned int& packet_offset,
+                                       unsigned int& downlink_frame_offset)
 {
-    const size_t field_size = field_bits.size();
+    const unsigned int field_size = field_bits.size();
     const int field_overflow = (field_size + packet_offset)
         - DownlinkProducer::num_bits_in_packet; // Number of bits in field that run past the packet end
 
@@ -119,9 +119,9 @@ void DownlinkProducer::execute() {
 
     char* snapshot_ptr = snapshot_ptr_f.get();
     // Create the required iterators
-    size_t downlink_frame_offset = 0; // Bit offset from the beginning
+    unsigned int downlink_frame_offset = 0; // Bit offset from the beginning
                                       // of the snapshot buffer
-    size_t packet_offset = 0;         // Bit offset from the beginning
+    unsigned int packet_offset = 0;         // Bit offset from the beginning
                                       // of the current downlink packet
 
     // Add initial packet header
@@ -195,7 +195,7 @@ const std::vector<DownlinkProducer::Flow>& DownlinkProducer::get_flows() const {
 
 DownlinkProducer::Flow::Flow(const StateFieldRegistry& r,
                         const FlowData& flow_data,
-                        const size_t num_flows) : id_sr(num_flows),
+                        const unsigned int num_flows) : id_sr(num_flows),
                                                   is_active(flow_data.is_active)
 {
     if (flow_data.id > num_flows || flow_data.id == 0) {
@@ -226,9 +226,9 @@ DownlinkProducer::Flow::Flow(const StateFieldRegistry& r,
     assert(get_packet_size() <= num_bits_in_packet - 1 - 32); // Flow should fit within one downlink packet
 }
 
-size_t DownlinkProducer::Flow::get_packet_size() const {
+unsigned int DownlinkProducer::Flow::get_packet_size() const {
     // Get bitcount of all fields in the flow
-    size_t packet_size = 0;
+    unsigned int packet_size = 0;
     packet_size += id_sr.bitsize();
 
     for(auto const& field: field_list) {
@@ -244,7 +244,7 @@ void DownlinkProducer::toggle_flow(unsigned char id) {
         assert(false);
     }
 
-    for(size_t idx = 0; idx < flows.size(); idx++) {
+    for(unsigned int idx = 0; idx < flows.size(); idx++) {
         unsigned char flow_id;
         flows[idx].id_sr.deserialize(&flow_id);
         if (flow_id == id) {
@@ -269,8 +269,8 @@ void DownlinkProducer::shift_flow_priorities(unsigned char id1, unsigned char id
         assert(false);
     }
 
-    size_t idx1 = 0, idx2 = 0;
-    for(size_t idx = 0; idx < flows.size(); idx++) {
+    unsigned int idx1 = 0, idx2 = 0;
+    for(unsigned int idx = 0; idx < flows.size(); idx++) {
         unsigned char flow_id;
         flows[idx].id_sr.deserialize(&flow_id);
         if (flow_id == id1) {
@@ -282,12 +282,12 @@ void DownlinkProducer::shift_flow_priorities(unsigned char id1, unsigned char id
     }
     
     if (idx1>idx2) {
-        for (size_t i = idx1; i > idx2; i--) {
+        for (unsigned int i = idx1; i > idx2; i--) {
             std::swap(flows[i], flows[i-1]);
         }
     }
     else if (idx2>idx1) {
-        for (size_t i = idx1; i < idx2; i++) {
+        for (unsigned int i = idx1; i < idx2; i++) {
             std::swap(flows[i],flows[i+1]);
         }
     }
