+++ src/fsw/FCCode/EEPROMControllerCommon.cpp
@@ -1,7 +1,7 @@
 #include "EEPROMController.hpp"
 
 void EEPROMController::init(const std::vector<std::string>& statefields, const std::vector<unsigned int>& periods){
-  for (size_t i = 0; i<statefields.size(); i++){
+  for (unsigned int i = 0; i<statefields.size(); i++){
     // copy the string name of the statefield into a char array
     char field[statefields.at(i).length() + 1];
     strcpy(field, statefields.at(i).c_str());
@@ -26,7 +26,7 @@ void EEPROMController::init(const std::vector<std::string>& statefields, const s
 
 void EEPROMController::execute() {
   //if enough control cycles have passed, write the field values to EEPROM
-  for (size_t i = 0; i<pointers.size(); i++) {
+  for (unsigned int i = 0; i<pointers.size(); i++) {
     if(control_cycle_count%sf_periods.at(i)==0){
       update_EEPROM(i);
     }
